+++ templates/vite-cloudflare/functions/[[path]].ts
@@ -5,7 +5,4 @@ import { createPagesFunctionHandler } from "@remix-run/cloudflare-pages";
 // eslint-disable-next-line import/no-unresolved
 import * as build from "../build/server";
 
-export const onRequest = createPagesFunctionHandler({
-  build,
-  getLoadContext: (context) => ({ env: context.env }),
-});
+export const onRequest = createPagesFunctionHandler({ build });